refactor(#716): single IoBinding/IoEndpoint + ChannelMode in domain (kill duplicate)
Consolidates two divergent definitions (one in project with DeviceId +
ChainInputMode, one in infra-filesystem with bare String device_id) into a
single canonical type in the domain crate.

- crates/domain/src/io_binding.rs: new home — ChannelMode enum +
  IoEndpoint (with DeviceId) + IoBinding; re-exported from domain::lib.
- crates/project/src/io_binding.rs: replaced with pub use re-export from
  domain; path project::io_binding::{IoBinding,IoEndpoint,ChannelMode}
  unchanged.
- crates/infra-filesystem: added domain dep; replaced local IoEndpoint/
  IoBinding structs with re-export; test updated to use DeviceId + mode.
