+++ crates/domain/src/lib.rs
@@ -1,5 +1,8 @@
 pub mod ids;
+pub mod io_binding;
 pub mod value_objects;
+
+pub use io_binding::{ChannelMode, IoBinding, IoEndpoint};
 
 #[cfg(test)]
 mod lib_tests;
