+++ crates/infra-filesystem/src/lib.rs
@@ -224,38 +224,9 @@ pub fn asset_paths() -> &'static AssetPaths {
 pub struct FilesystemStorage;
 
 // ── I/O binding registry (#716) ────────────────────────────────────────────
-
-/// A single named endpoint (input or output channel group) on a physical device.
-///
-/// Stored in the per-machine system config (`config.yaml`) so that projects
-/// can reference endpoints by name without hardcoding device paths (ADR 0003).
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct IoEndpoint {
-    /// Human-readable label (e.g. `"Guitar In 1"`).
-    pub name: String,
-    /// Opaque stable identifier of the physical device that owns this endpoint.
-    pub device_id: String,
-    /// Zero-based channel indices on the device.
-    pub channels: Vec<usize>,
-}
-
-/// A complete I/O binding: a named group of input + output endpoints, identified
-/// by a stable `id`.
-///
-/// Stored in [`AppConfig::io_bindings`] (per-machine registry). Projects
-/// reference bindings by `id`, not by device path, so `.openrig` files stay
-/// portable across machines.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct IoBinding {
-    /// Stable registry key (e.g. `"main"`, `"monitor"`).
-    pub id: String,
-    /// Human-readable display name (e.g. `"Scarlett 2i2"`).
-    pub name: String,
-    /// Input endpoints exposed by this binding.
-    pub inputs: Vec<IoEndpoint>,
-    /// Output endpoints exposed by this binding.
-    pub outputs: Vec<IoEndpoint>,
-}
+// Single source of truth lives in `domain::io_binding`; re-exported here so
+// `AppConfig::io_bindings` uses the canonical types.
+pub use domain::io_binding::{ChannelMode, IoBinding, IoEndpoint};
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
 pub struct RecentProjectEntry {
